Rename classes.py to datatypes.py
