+++ examples/dalle-bot/main.py
@@ -1,7 +1,7 @@
 from textbase import bot, Message
 from textbase.models import DallE
 from typing import List
-from textbase.classes import Image
+from textbase.datatypes import Image
 
 # Load your OpenAI API key
 DallE.api_key = ""
