+++ textbase/datatypes.py
@@ -1,4 +1,4 @@
-from textbase.helpers import convert_img_to_url
+from textbase.helpers import convert_img_to_url, convert_video_to_url, convert_audio_to_url
 from PIL.Image import Image as PILImageClass
 import os
 
@@ -23,3 +23,27 @@ class Image:
 
     def upload_file_to_bucket(self):
         self.url = convert_img_to_url(image_file_path=self.path)
+
+class Video:
+    def __init__(self, url="", path=""):
+        self.url = url
+
+        if path and not os.path.exists(path):
+            raise FileNotFoundError("The given path doesn't exist.")
+
+        self.path = path
+
+    def upload_file_to_bucket(self):
+        self.url = convert_video_to_url(video_file_path=self.path)
+
+class Audio:
+    def __init__(self, url="", path=""):
+        self.url = url
+
+        if path and not os.path.exists(path):
+            raise FileNotFoundError("The given path doesn't exist.")
+
+        self.path = path
+
+    def upload_file_to_bucket(self):
+        self.url = convert_audio_to_url(audio_file_path=self.path)
